Refactor examples/tpm_io into separate platform source files
* tpm_io_atmel.c contains only ASF code
* tpm_io_barebox.c contains only code for Barebox
* tpm_io_st.c to contain only STM32 specific code
* tpm_io_linux.c to contain only Linux specific code
* tpm_io_qnx includes the new QNX callback
* tpm_io_xilinx.c to contain only Xilinx code

- Noted that native Linux /dev/tpmX does not require an IO callback
- TPM Simulator and Windows TBS also does not require an IO callback
